move op_close to cli/ops/resources.rs
